--- components/Banner/index.ts
@@ -1,2 +1,0 @@
-export type { BannerProps } from './Banner';
-export { default as Banner } from './Banner';
