--- components/Button/index.ts
@@ -1,2 +1,0 @@
-export type { ButtonProps } from './Button';
-export { default as Button } from './Button';
