+++ components/Collection/EditCollectionModal.tsx
@@ -1,5 +1,5 @@
-import { Button } from '@components/Button';
 import { Modal } from '@components/Modal';
+import Button from '@components/UI/Button';
 import { useEditCollectionMutation } from '@graphql/collection.graphql';
 import { useDeleteMutation } from '@graphql/post.graphql';
 import { Switch } from '@headlessui/react';
