+++ components/PostCard/PostCard.tsx
@@ -1,6 +1,6 @@
-import { Button } from '@components/Button';
 import { CommentHandler } from '@components/Comment';
 import { LoginContext } from '@components/Layout/Layout';
+import Button from '@components/UI/Button';
 import {
   useCreateCommentMutation,
   useDeleteMutation,
