+++ pages/register.tsx
@@ -1,6 +1,6 @@
-import { Button } from '@components/Button/';
 import { Input } from '@components/Input';
 import { Layout } from '@components/Layout';
+import Button from '@components/UI/Button';
 import { useMeQuery } from '@graphql/user.graphql';
 import { useCreateUserMutation } from '@graphql/user.graphql';
 import Link from 'next/link';
